cleanup email tests, move to new directory, split up into three files - move setting some default config settings into conftest.py so they apply to all tests
